Changing entities

+++ DatabaseModel/JobTitleEntity.cs
@@ -5,7 +5,7 @@
         public int Id { get; set; }
         public required string Name { get; set; }
         public string? DutiesDescription { get; set; }
-        public int BaseSalary { get; set; }
+        public double BaseSalary { get; set; }
         public virtual ICollection<WarehouseEmployeeEntity> WarehouseEmployees { get; set; } = [];
     }
 }
